pipeline/requirements.txt
psycopg2-binary==2.9.9
redis==5.0.1
python-dateutil==2.8.2